#27 - optimize new images from server as they're added
