+++ bin/build.js
@@ -1,40 +1,17 @@
-const imagemin = require('imagemin');
-const imageminSvgo = require('imagemin-svgo');
+const optimize = require('./optimize');
 const generatePathFile = require('./path-data');
-let options = {
-  plugins: [
-    {cleanupIDs: {remove: false}},
-    {removeStyleElement: true},
-    {removeUselessDefs: true},
-    {removeUselessStrokeAndFill: true},
-    {removeHiddenElems: true},
-    {removeEmptyText: true},
-    {convertShapeToPath: true},
-    {removeEmptyAttrs: true},
-    {removeEmptyContainers: true},
-    {mergePaths: true},
-    {removeTitle: true},
-    {removeDesc: true},
-    {removeDimensions: true},
-    {removeAttrs: {attrs: ['class', '(stroke|fill)']}}
-  ],
-  multipass: true
-};
 
-module.exports = function (log) {
-  return imagemin(['*.svg'], './', { use: [imageminSvgo(options)] })
+module.exports = function () {
+  return optimize(['*.svg'], './')
     .then(function (result) {
-      options.plugins[0] = {cleanupIDs: {remove: true}};
-      return imagemin(['icons/*.svg'], 'icons/', { use: [imageminSvgo(options)] });
+      return optimize(['icons/*.svg'], 'icons/', true);
     })
     .catch(error => {
       console.error('🚨  Error while optimizing icons');
       throw error;
     })
     .then(function (result) {
-      if (log) {
-        console.log('✨  icons optimized successfully');
-      }
+      console.log('✨  icons optimized successfully');
       return generatePathFile();
     })
     .catch(error => {
@@ -42,9 +19,7 @@ module.exports = function (log) {
       throw error;
     })
     .then(function (files) {
-      if (log) {
-        console.log('✨  path file generated at ./docs/icons.json');
-      }
+      console.log('✨  path file generated at ./docs/icons.json');
       return files;
     });
 }
