+++ bin/server.js
@@ -1,29 +1,42 @@
 #!/usr/bin/env node
 const build = require('./build');
 const pathData = require('./path-data');
+const optimize = require('./optimize');
+const debounce = require('debounce');
 const bs = require('browser-sync').create();
 const fs = require('fs');
+const options = {
+  awaitWriteFinish: true,
+  ignoreInitial: true
+};
 
+console.log('🗜  optimizing icons...')
 
-console.log('optimizing icons...')
+build()
+  .then(() => {
+    bs.init({
+      server: './docs',
+      notify: false,
+      ui: false,
+      port: 8080
+    });
+    bs.watch('icons/*.svg', options, onChange);
+    bs.watch('docs/keywords.json', options, onChange);
 
-build(true).then(() => {
-  bs.init({
-    server: './docs',
-    notify: false,
-    ui: false,
-    port: 8080
-  });
-  bs.watch('icons/*.svg', {ignoreInitial: true}, onChange);
-  bs.watch('docs/keywords.json', {ignoreInitial: true}, onChange);
-
-  function onChange (event, file) {
-    if (event === 'change' || event === 'add') {
-      console.log('icon changes detected, rebuilding...')
-      pathData().then(files => {
-        console.log('✨  path file updated');
-        bs.reload();
-      });
+    function onChange (event, file) {
+      if (event === 'add') {
+        console.log('🗜  new icon detected, optimizing...')
+        optimize([file], 'icons/', true).then(() => {});
+      } else {
+        update();
+      }
     }
-  }
-});
+
+    const update = debounce(function () {
+      pathData()
+        .then(files => {
+          console.log('✨  path file updated');
+          bs.reload();
+        });
+    }, 300);
+  });
